# api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/me": {
      "get": {
        "operationId": "MeController_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeWithMembershipsDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED: no valid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED: too many requests, retry after the Retry-After header's seconds",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get the current user and their location memberships (for onboarding routing)",
        "tags": [
          "Me"
        ]
      },
      "patch": {
        "operationId": "MeController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeDto"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR: the request is invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED: no valid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED: too many requests, retry after the Retry-After header's seconds",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Change the current user's email language (ADR 0018)",
        "tags": [
          "Me"
        ]
      },
      "delete": {
        "operationId": "MeController_remove",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "UNAUTHENTICATED: no valid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "REAUTHENTICATION_REQUIRED: the session is too old, sign in again first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "CONFLICT: not possible in the resource's current state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED: too many requests, retry after the Retry-After header's seconds",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Delete the current user's account, needs a sign-in in the last 10 minutes (ADR 0018)",
        "tags": [
          "Me"
        ]
      }
    },
...
    "/locations/{locationId}/memberships": {
      "get": {
        "operationId": "MembershipController_list",
        "parameters": [
          {
            "name": "locationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MemberDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED: no valid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "UNAUTHORIZED: signed in, but not allowed to do this",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND: the resource doesn't exist, or isn't visible to you",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED: too many requests, retry after the Retry-After header's seconds",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "List a location's members",
        "tags": [
          "Memberships"
        ]
      }
    },
    "/locations/{locationId}/memberships/me": {
      "delete": {
        "operationId": "MembershipController_leave",
        "parameters": [
          {
            "name": "locationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "UNAUTHENTICATED: no valid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "UNAUTHORIZED: signed in, but not allowed to do this",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND: the resource doesn't exist, or isn't visible to you",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "CONFLICT: not possible in the resource's current state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED: too many requests, retry after the Retry-After header's seconds",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Leave this location, removing your own membership (not the owner, ADR 0017)",
        "tags": [
          "Memberships"
        ]
      }
    },
    "/locations/{locationId}/memberships/{membershipId}": {
      "patch": {
        "operationId": "MembershipController_updateRole",
        "parameters": [
          {
            "name": "locationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "membershipId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMembershipRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberDto"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR: the request is invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED: no valid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "UNAUTHORIZED: signed in, but not allowed to do this",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND: the resource doesn't exist, or isn't visible to you",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "CONFLICT: not possible in the resource's current state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED: too many requests, retry after the Retry-After header's seconds",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Change a member's role between ADMIN and EMPLOYEE (owner only, ADR 0017)",
        "tags": [
          "Memberships"
        ]
      },
      "delete": {
        "operationId": "MembershipController_remove",
        "parameters": [
          {
            "name": "locationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "membershipId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "UNAUTHENTICATED: no valid session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "UNAUTHORIZED: signed in, but not allowed to do this",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "NOT_FOUND: the resource doesn't exist, or isn't visible to you",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "CONFLICT: not possible in the resource's current state",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "RATE_LIMITED: too many requests, retry after the Retry-After header's seconds",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Remove a member's access (not the owner, an admin removes employees only)",
        "tags": [
          "Memberships"
        ]
      }
    },
...
  },
  "info": {
    "title": "ReadyYet API",
    "description": "Better Auth's own routes (/api/auth/*) aren't included here, they're raw middleware, not Nest controllers. See ADR 0008/0011.",
    "version": "0.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token"
      }
    },
    "schemas": {
      "ValidationIssueDto": {
        "type": "object",
        "properties": {
          "constraints": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          },
          "property": {
            "type": "string"
          }
        },
        "required": [
          "property"
        ]
      },
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "code": {
            "enum": [
              "VALIDATION_ERROR",
              "UNAUTHENTICATED",
              "UNAUTHORIZED",
              "REAUTHENTICATION_REQUIRED",
              "NOT_FOUND",
              "CONFLICT",
              "RATE_LIMITED",
              "INTERNAL_ERROR"
            ],
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationIssueDto"
            }
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ApiErrorResponseDto": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorDto"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "MyBusinessDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "MyLocationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "business": {
            "$ref": "#/components/schemas/MyBusinessDto"
          }
        },
        "required": [
          "id",
          "name",
          "business"
        ]
      },
      "MyMembershipDto": {
        "type": "object",
        "properties": {
          "role": {
            "enum": [
              "OWNER",
              "ADMIN",
              "EMPLOYEE"
            ],
            "type": "string"
          },
          "location": {
            "$ref": "#/components/schemas/MyLocationDto"
          }
        },
        "required": [
          "role",
          "location"
        ]
      },
      "MeWithMembershipsDto": {
        "type": "object",
        "properties": {
          "locale": {
            "enum": [
              "EN",
              "FR"
            ],
            "type": "string"
          },
          "memberships": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MyMembershipDto"
            }
          },
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "locale",
          "memberships",
          "id",
          "email",
          "name"
        ]
      },
      "UpdateMeDto": {
        "type": "object",
        "properties": {
          "locale": {
            "type": "string",
            "enum": [
              "EN",
              "FR"
            ]
          }
        },
        "required": [
          "locale"
        ]
      },
      "MeDto": {
        "type": "object",
        "properties": {
          "locale": {
            "enum": [
              "EN",
              "FR"
            ],
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "locale",
          "id",
          "email",
          "name"
        ]
      },
...
      "MemberUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "name"
        ]
      },
      "MemberDto": {
        "type": "object",
        "properties": {
          "role": {
            "enum": [
              "OWNER",
              "ADMIN",
              "EMPLOYEE"
            ],
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "locationId": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/MemberUserDto"
          }
        },
        "required": [
          "role",
          "id",
          "userId",
          "locationId",
          "createdAt",
          "user"
        ]
      },
      "UpdateMembershipRoleDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "EMPLOYEE"
            ]
          }
        },
        "required": [
          "role"
        ]
      },
...
    }
  }
}